procure_iq_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
requests
python-dotenv
httpx
